Add bootstrapped Blueprint Architecture sample app
The bootstrap module was added as an example
and tool for quickly beginning a blueprint project.

The bootstrapped application is setup with modern
tools for app building, including dagger 2.0
dependency injection, rxjava and retrofit for
api requests. Clients can use the bootstrap template
as a starter to avoid re-writing all the boilerplate
of a blueprint app.