Add Support for library project and AGP 5
In AGP 5 (and currently for AARs), resource IDs are no
longer final, which means we can't use them in annotation
parameters the way we do with ComponentViewHolder

This commit changes the ComponentViewHolder annotation to
accept a String, which must be the name of the layout for the
view holder, so that blueprint can continue to function. Blueprint
will use the name of the layout to fabricate the layout id, and use
that the same way as before, for component view inflation and referencing.